Early draft with 16 (!) manifesto points
Mostly placeholder text throughout. Also introduces an even more eye-
searing version of the CRT/green-screen effect, as well as the
MooTools library. jQuery was just barely on the radar at that point,
or at least my radar. So I went with MooTools. So did, in turn, the
Kairos Toolbar.