The Lo-Fi Manifesto, as originally submitted to Kairos
I actually had to triangulate this version with reviewer feedback
to make sure that it was, in fact, the as-submitted version. As
best as I can tell, it is.

The writing is pretty clunky, even in its better moments. The
individual manifesto points are pretty spare--a point that one
reviewer remarked on.

And of course, the visual design is different, relying on a black
background with an alien greenish glow at top and bottom of the
expanded manifesto points.

The first manifesto point is also already open when the screen
loads; hard to say whether I thought that would make the webtext
a little more user friendly. I was probably more concerned that
the reviewers would miss that there was content hidden on the page.
Either way, one of the reviewers suggested collapsing all of the
points, which is of course how it ultimately appeared.